# public/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Piyachok API",
    "description": "Документация backend-платформы проекта Piyachok."
  },
  "host": "localhost:3000",
  "basePath": "/api",
...
  "schemes": ["http"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Введіть токен у форматі: Bearer <AccessToken>"
    }
  },
  "paths": {
...
    "/tags": {
      "get": {
        "tags": ["Tags"],
        "summary": "Отримання списку всіх тегів",
        "description": "Повертає повний список тегів, відсортованих за алфавітом (A-Z, А-Я). Доступно публічно.",
        "responses": {
          "200": {
            "description": "Список тегів успішно отримано",
            "schema": {
              "type": "object",
              "properties": {
                "tags": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "example": "tag-111-uuid" },
                      "name": { "type": "string", "example": "Крафт" }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      },
      "post": {
        "tags": ["Tags"],
        "summary": "Створення нового тегу",
        "description": "Дозволяє авторизованому користувачу створити новий тег. Назва тегу має бути унікальною.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["name"],
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Жива музика",
                  "description": "Назва тегу (буде автоматично очищена від зайвих пробілів)"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Тег успішно створено",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Тег успішно створено" },
                "tag": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "example": "tag-222-uuid" },
                    "name": { "type": "string", "example": "Жива музика" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Не передано назву тегу або такий тег вже існує в системі",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "401": {
            "description": "Неавторизований доступ",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      }
    },
...
    "/venues": {
      "get": {
        "tags": ["Venues"],
        "summary": "Отримання списку закладу з фільтрацією та пагінацією",
        "description": "Повертає відфільтрований та відсортований список закладів разом з усім переліком доступних тегів для зручної побудови фільтрів на фронтенді. Підтримує сортування за відстанню (гео-координатами).",
        "parameters": [
          { "name": "page", "in": "query", "type": "integer", "default": 1, "description": "Номер сторінки" },
          { "name": "limit", "in": "query", "type": "integer", "default": 9, "description": "Елементів на сторінці" },
          { "name": "search", "in": "query", "type": "string", "description": "Пошук за назвою або тегами" },
          { "name": "type", "in": "query", "type": "string", "description": "Тип закладу (PUB, RESTAURANT тощо)" },
          { "name": "minCheck", "in": "query", "type": "integer", "description": "Мінімальний середній чек" },
          { "name": "maxCheck", "in": "query", "type": "integer", "description": "Максимальний середній чек" },
          { "name": "tags", "in": "query", "type": "string", "description": "Теги через кому (наприклад: Крафт,Тераса)" },
          { "name": "hasWifi", "in": "query", "type": "boolean", "description": "Фільтр наявності Wi-Fi" },
          { "name": "hasParking", "in": "query", "type": "boolean", "description": "Фільтр наявності парковки" },
          { "name": "hasLiveMusic", "in": "query", "type": "boolean", "description": "Фільтр живої музики" },
          { "name": "sortBy", "in": "query", "type": "string", "enum": ["createdAt", "rating", "averageCheck", "name", "distance"], "default": "createdAt" },
          { "name": "order", "in": "query", "type": "string", "enum": ["asc", "desc"], "default": "desc" },
          { "name": "lat", "in": "query", "type": "number", "description": "Широта поточного користувача (потрібна для сортування за distance)" },
          { "name": "lng", "in": "query", "type": "number", "description": "Довгота поточного користувача (потрібна для сортування за distance)" },
          { "name": "myOwn", "in": "query", "type": "boolean", "description": "Тільки для MANAGER/ADMIN: отримати суто мої заклади" }
        ],
        "responses": {
          "200": {
            "description": "Успішно завантажено список закладів",
            "schema": {
              "type": "object",
              "properties": {
                "venues": {
                  "type": "array",
                  "items": { "type": "object" }
                },
                "allTags": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string" },
                      "name": { "type": "string" }
                    }
                  }
                },
                "pagination": {
                  "type": "object",
                  "properties": {
                    "currentPage": { "type": "integer" },
                    "totalPages": { "type": "integer" },
                    "totalItems": { "type": "integer" },
                    "hasNextPage": { "type": "boolean" },
                    "hasPrevPage": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      },
      "post": {
        "tags": ["Venues"],
        "summary": "Створення нового закладу",
        "description": "Дозволяє користувачам з ролями MANAGER та SUPERADMIN додати новий заклад у каталог. Заклади від менеджерів потрапляють на модерацію.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["name", "type", "mainImage", "address", "workingHours", "averageCheck"],
              "properties": {
                "name": { "type": "string", "example": "П'яний Лис" },
                "type": { "type": "string", "example": "PUB" },
                "mainImage": { "type": "string", "example": "/uploads/temp-fox.jpg", "description": "Тимчасовий шлях до фото" },
                "images": { "type": "array", "items": { "type": "string" }, "example": ["/uploads/temp-interior.jpg"] },
                "address": { "type": "string", "example": "вулиця Садова, 12" },
                "workingHours": { "type": "string", "example": "10:00 - 23:00" },
                "averageCheck": { "type": "integer", "example": 350 },
                "description": { "type": "string", "example": "Затишний паб з великим вибором пива..." },
                "phone": { "type": "string", "example": "+380671234567" },
                "tags": { "type": "array", "items": { "type": "string" }, "example": ["Крафт", "Бургери"] },
                "latitude": { "type": "number", "example": 46.4825 },
                "longitude": { "type": "number", "example": 30.7233 },
                "hasWifi": { "type": "boolean", "example": true },
                "hasParking": { "type": "boolean", "example": true },
                "hasLiveMusic": { "type": "boolean", "example": false }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Заклад успішно додано",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Заклад успішно створено та відправлено на модерацію до Адміністратора." },
                "venue": { "type": "object" }
              }
            }
          },
          "400": {
            "description": "Не заповнено обов'язкові поля",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "401": {
            "description": "Користувач не авторизований",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "403": {
            "description": "У вас немає ролі MANAGER або SUPERADMIN",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      }
    }
  },
  "definitions": {
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "message": { "type": "string", "example": "Внутрішня помилка сервера / Доступ заборонено" }
      }
    },
...
  }
}